CM-911: Move the project into another subfolder
