--- csharp-intermediate-topics/InMemoryZipFilesInNet/InMemoryZipFilesInNet.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InMemoryZipFilesInNet", "InMemoryZipFilesInNet\InMemoryZipFilesInNet.csproj", "{EEE49998-F149-4788-A12B-A229646F20FA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{8131EA0E-6B84-400D-BF74-49AF8A56C93B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EEE49998-F149-4788-A12B-A229646F20FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EEE49998-F149-4788-A12B-A229646F20FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EEE49998-F149-4788-A12B-A229646F20FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EEE49998-F149-4788-A12B-A229646F20FA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8131EA0E-6B84-400D-BF74-49AF8A56C93B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8131EA0E-6B84-400D-BF74-49AF8A56C93B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8131EA0E-6B84-400D-BF74-49AF8A56C93B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8131EA0E-6B84-400D-BF74-49AF8A56C93B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
