+++ files-csharp/InMemoryZipFilesInNet/InMemoryZipFilesInNet/IGetFile.cs
