+++ files-csharp/InMemoryZipFilesInNet/InMemoryZipFilesInNet/InMemoryZipFilesInNet.http
@@ -1,0 +1,21 @@
+@InMemoryZipFilesInNet_HostAddress = http://localhost:5293
+
+GET {{InMemoryZipFilesInNet_HostAddress}}/test-file/
+Accept: application/zip
+
+###
+
+GET {{InMemoryZipFilesInNet_HostAddress}}/create-in-memory-zip-file/
+Accept: application/zip
+
+###
+
+GET {{InMemoryZipFilesInNet_HostAddress}}/create-in-memory-zip-file-as-byte-array/
+Accept: application/zip
+
+###
+
+GET {{InMemoryZipFilesInNet_HostAddress}}/downloading-bigger-file/
+Accept: application/zip
+
+###
