+++ files-csharp/InMemoryZipFilesInNet/InMemoryZipFilesInNet/Services/FileService.cs
